Change files to ts
